move instead of unlink + phantom bug fixe

+++ server/poster/cleaner.php
@@ -6,9 +6,14 @@
 
   $unlinkeds = [];
   foreach ($_GET['items'] as $value) {
-    $u = unlink($value);
-    if($u)
-      $unlinkeds[] = $value; 
+    $filename = basename($value);
+    
+    $c = copy($value, '../assets/02-archives/'.$filename);
+    if($c){
+      $u = unlink($value);
+      if($u)
+        $unlinkeds[] = $value;  
+    }
   }
 
   $rep['unlinkeds'] = $unlinkeds;
